Fix multiple copies of elements with the same ids generated by macros.
If there is a -- in a slide, the content is copied, including elements
with a defined id.

Therefore, we now use classes to identify elements.

As a nice consequence, we can now share the container elements for
images, PDFs and DOTs with the same source, allowing to perform AJAX
requests and renderings only once.