Fix 02.md mistake in MLE formula (wrong signs)
I suspect an error in the slides: Most of the minus signs should actually be + signs. In the book (Eq. 5.64, 5.65) there is no minus in front of the probability, but there is in the slides, hence there are double negatives and the signs should be positive. :-) It can be verified manually on a paper. Plus, it would not make sense to argmin -MSE, because that would result in argmax MSE, which is incorrect. I hope I did not make any mistake, I edited the .md manually.